+++ examples/complete/main.tf
@@ -16,7 +16,7 @@ module "cloudfront_distribution" {
   }
 
   # By-default, fqdn for the CDN should be added, it should be the one for which certificate is issued
-  domain_aliases      = ["example.example.com"]
+  domain_aliases      = ["example.example.com", "example1.example.com"]
   default_root_object = ""
 
   # Default behavior
@@ -34,14 +34,14 @@ module "cloudfront_distribution" {
 
 
   # DNS Mapping variables
-  cdn_certificate_arn        = module.acm_virginia.certificate_arn[0]
-  acm_cert_domain_name       = "example.example.com"
-  route53_domain_name        = "example.com"
-  is_enable_waf              = true
-  is_enable_waf_default_rule = false
-  waf_default_action         = "allow"
-  tags                       = var.custom_tags
+  is_automatic_create_dns_record = true
+  cdn_certificate_arn            = module.acm_virginia.certificate_arn[0]
+  route53_domain_name            = "example.com"
+  is_enable_waf                  = true
+  is_enable_waf_default_rule     = false
+  waf_default_action             = "allow"
+  tags                           = var.custom_tags
   providers = {
     aws = aws.virginia
-   }
+  }
 }